pyproject.toml
[tool.poetry]
name = "authservice"
version = "0.1.0"
description = ""
authors = ["denisahmetov <denis.akhmetov.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
asyncpg = "^0.28.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
greenlet = "^3.0.0"
pydantic = {extras = ["email"], version = "^2.4.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
